Move package of steps
