+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/steps/GetPorts.kt
@@ -1,7 +1,7 @@
 // Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.services.lambda.upload.steps
+package software.aws.toolkits.jetbrains.services.lambda.steps
 
 import com.intellij.util.net.NetUtils
 import software.aws.toolkits.core.utils.AttributeBagKey
